backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
bcrypt
PyJWT
pydantic
qrcode
python-multipart
